+++ src/Connector/Recoverable/ExponentialSleepRecoverableExceptionHandler.php
@@ -7,19 +7,22 @@ use Amp\Promise;
 use ScriptFUSION\Retry\ExceptionHandler\ExponentialBackoffExceptionHandler;
 
 /**
- * Sleeps for an exponentially increasing series of delays specified in microseconds.
+ * Sleeps for an exponentially increasing series of delays.
  */
 class ExponentialSleepRecoverableExceptionHandler implements RecoverableExceptionHandler
 {
     private $initialDelay;
 
+    /**
+     * @var ExponentialBackoffExceptionHandler
+     */
     private $handler;
 
     /**
      * Initializes this instance with the specified initial delay. The initial delay will be used when the first
      * exception is handled; subsequent exceptions will cause longer delays.
      *
-     * @param int $initialDelay Initial delay.
+     * @param int $initialDelay Initial delay in microseconds.
      */
     public function __construct(int $initialDelay = ExponentialBackoffExceptionHandler::DEFAULT_COEFFICIENT)
     {
@@ -33,6 +36,8 @@ class ExponentialSleepRecoverableExceptionHandler implements RecoverableExceptio
 
     public function __invoke(RecoverableException $exception): ?Promise
     {
-        return ($this->handler)($exception);
+        ($this->handler)();
+
+        return null;
     }
 }
