+++ src/Specification/IncompatibleTransformerException.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace ScriptFUSION\Porter\Specification;
+
+/**
+ * The exception that is thrown when the wrong type of transformer is added to a specification.
+ */
+class IncompatibleTransformerException extends \LogicException
+{
+    // Intentionally empty.
+}
