+++ src/Transform/AnysyncTransformer.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace ScriptFUSION\Porter\Transform;
+
+/**
+ * Marker interface that specifies a transformer that is synchronous, asynchronous or both.
+ */
+interface AnysyncTransformer
+{
+    // Marker interface.
+}
