+++ src/Transform/Transformer.php
@@ -8,7 +8,7 @@ use ScriptFUSION\Porter\Collection\RecordCollection;
 /**
  * Provides a method to transform imported data.
  */
-interface Transformer
+interface Transformer extends AnysyncTransformer
 {
     /**
      * Transforms the specified record collection decorated with the specified context data.
