+++ test/Unit/Porter/ImportSpecificationTest.php
@@ -8,6 +8,8 @@ use ScriptFUSION\Porter\Connector\Recoverable\RecoverableExceptionHandler;
 use ScriptFUSION\Porter\Provider\Resource\ProviderResource;
 use ScriptFUSION\Porter\Specification\DuplicateTransformerException;
 use ScriptFUSION\Porter\Specification\ImportSpecification;
+use ScriptFUSION\Porter\Specification\IncompatibleTransformerException;
+use ScriptFUSION\Porter\Transform\AsyncTransformer;
 use ScriptFUSION\Porter\Transform\Transformer;
 
 /**
@@ -78,6 +80,9 @@ final class ImportSpecificationTest extends TestCase
         self::assertCount(2, $this->specification->getTransformers());
         self::assertContains($transformer1, $this->specification->getTransformers());
         self::assertContains($transformer2, $this->specification->getTransformers());
+
+        $this->expectException(IncompatibleTransformerException::class);
+        $this->specification->addTransformer(\Mockery::mock(AsyncTransformer::class));
     }
 
     public function testAddTransformers(): void
